# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/emaillist/blacklist": {
            "get": {
                "description": "获取用户的黑名单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "获取黑名单",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "添加邮箱地址或域名到黑名单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "添加到黑名单",
                "parameters": [
                    {
                        "description": "黑名单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.AddToBlacklistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/emaillist/blacklist/{id}": {
            "delete": {
                "description": "从黑名单中删除指定条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "从黑名单中删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/emaillist/whitelist": {
            "get": {
                "description": "获取用户的白名单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "获取白名单",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "添加邮箱地址或域名到白名单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "添加到白名单",
                "parameters": [
                    {
                        "description": "白名单信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.AddToWhitelistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/emaillist/whitelist/{id}": {
            "delete": {
                "description": "从白名单中删除指定条目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emaillist"
                ],
                "summary": "从白名单中删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "条目 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.AddToBlacklistRequest": {
            "type": "object",
            "required": [
                "target"
            ],
            "properties": {
                "reason": {
                    "description": "添加原因",
                    "type": "string"
                },
                "target": {
                    "description": "邮箱地址或域名",
                    "type": "string"
                }
            }
        },
        "internal_handler.AddToWhitelistRequest": {
            "type": "object",
            "required": [
                "target"
            ],
            "properties": {
                "reason": {
                    "description": "添加原因",
                    "type": "string"
                },
                "target": {
                    "description": "邮箱地址或域名",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}